seprate CongratulationsCard from claim card
