+++ src/components/cards/Claim.tsx
@@ -1,6 +1,5 @@
 import { FC, useContext, useEffect, useState } from 'react';
 import Image from 'next/image';
-import Link from 'next/link';
 import styled from 'styled-components';
 import { Button } from '../styled-components/Button';
 import { Row } from '../styled-components/Grid';
@@ -20,99 +19,6 @@ import type { TransactionResponse } from '@ethersproject/providers';
 import { wrongWallet } from '../toasts/claim';
 import { useTokenDistro } from '@/context/tokenDistro.context';
 import { H2, Lead } from '@giveth/ui-design-system';
-import SparkleBurstAnimation from '../../animations/sparkle-burst.json';
-import SparkleAnimation from '../../animations/sparkle.json';
-import Lottie from 'react-lottie';
-
-const ClaimedContainer = styled.div`
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	position: relative;
-`;
-
-const SunImage = styled.div`
-	position: absolute;
-	height: 0px;
-	left: 146px;
-	top: 211px;
-	@media only screen and (max-width: 1360px) {
-		left: 71px;
-		top: 180px;
-	}
-	@media only screen and (max-width: 1120px) {
-		display: none;
-	}
-`;
-
-const ClaimedTitle = styled.div`
-	padding-top: 60px;
-	font-family: 'Red Hat Text';
-	font-size: 64px;
-	font-weight: 700;
-	text-align: center;
-	@media only screen and (max-width: 1360px) {
-		padding-top: 30px;
-	}
-`;
-
-const ClaimedSubtitleContainer = styled.div`
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-	gap: 4px;
-	padding-left: 60px;
-	@media only screen and (max-width: 1360px) {
-		padding-left: 0px;
-	}
-`;
-
-const ClaimedSubtitleA = styled.div`
-	font-family: 'Red Hat Text';
-	font-size: 21px;
-	text-align: center;
-	display: flex;
-	gap: 12px;
-`;
-
-const AddGivButton = styled.div`
-	cursor: pointer;
-`;
-
-const ClaimedSubtitleB = styled.div`
-	font-family: 'Red Hat Text';
-	font-size: 20px;
-	text-align: center;
-`;
-
-const SocialButton = styled(Button)`
-	font-family: 'Red Hat Text';
-	font-size: 14px;
-	font-weight: bold;
-	text-transform: uppercase;
-	background-color: transparent;
-	border: 2px solid white;
-	height: 50px;
-	width: 265px;
-	margin-top: 12px;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	gap: 4px;
-`;
-
-const ExploreButton = styled(SocialButton)`
-	background-color: #e1458d;
-	border: none;
-	width: 285px;
-`;
-
-const ClaimFromAnother = styled.span`
-	cursor: pointer;
-	color: '#FED670'
-	margin-top: 4px;
-`;
 
 interface IClaimCardContainer {
 	claimed: any;
@@ -157,49 +63,10 @@ const MetamaskButton = styled.a`
 	cursor: pointer;
 `;
 
-const SparkleContainer = styled.div`
-	position: absolute;
-	right: 232px;
-	top: 73px;
-	@media only screen and (max-width: 1360px) {
-		right: 146px;
-	}
-	@media only screen and (max-width: 1120px) {
-		right: 36px;
-	}
-`;
-
-const SparkleBurstContainer = styled.div`
-	position: absolute;
-	left: 54px;
-	top: 70px;
-	@media only screen and (max-width: 1360px) {
-	}
-	@media only screen and (max-width: 1120px) {
-		display: none;
-	}
-`;
-
-const SparkleAnimationOptions = {
-	loop: false,
-	animationData: SparkleAnimation,
-	rendererSettings: {
-		preserveAspectRatio: 'xMidYMid slice',
-	},
-};
-
-const SparkleBurstAnimationOptions = {
-	loop: false,
-	animationData: SparkleBurstAnimation,
-	rendererSettings: {
-		preserveAspectRatio: 'xMidYMid slice',
-	},
-};
-
 const ClaimCard: FC<IClaimViewCardProps> = ({ index }) => {
-	const { activeIndex, goFirstStep, goPreviousStep } =
+	const { activeIndex, goPreviousStep, goNextStep } =
 		useContext(ClaimViewContext);
-	const { userAddress, totalAmount, resetWallet } = useUser();
+	const { userAddress, totalAmount } = useUser();
 	const {
 		isReady,
 		changeWallet,
@@ -213,22 +80,6 @@ const ClaimCard: FC<IClaimViewCardProps> = ({ index }) => {
 	const [txStatus, setTxStatus] = useState<TransactionResponse | undefined>();
 	const [showModal, setShowModal] = useState<boolean>(false);
 	const [showClaimModal, setShowClaimModal] = useState<boolean>(false);
-	const [isClaimed, setIsClaimed] = useState(false);
-	const [streamValue, setStreamValue] = useState<string>('0');
-
-	const { tokenDistroHelper } = useTokenDistro();
-
-	useEffect(() => {
-		setStreamValue(
-			formatWeiHelper(
-				tokenDistroHelper.getStreamPartTokenPerWeek(totalAmount),
-			),
-		);
-	}, [totalAmount, tokenDistroHelper]);
-
-	useEffect(() => {
-		setIsClaimed(false);
-	}, [address, userAddress]);
 
 	useEffect(() => {
 		setShowModal(
@@ -268,8 +119,8 @@ const ClaimCard: FC<IClaimViewCardProps> = ({ index }) => {
 	};
 
 	const onSuccess = (tx: TransactionResponse) => {
-		setIsClaimed(true);
 		setTxStatus(tx);
+		goNextStep();
 	};
 
 	return (
@@ -279,163 +130,37 @@ const ClaimCard: FC<IClaimViewCardProps> = ({ index }) => {
 				index={index}
 				claimed={txStatus}
 			>
-				{isClaimed ? (
-					<>
-						<SparkleBurstContainer>
-							<Lottie
-								options={SparkleBurstAnimationOptions}
-								height={200}
-								width={200}
-								isPaused={showClaimModal}
-								speed={0.8}
-							/>
-						</SparkleBurstContainer>
-						<SparkleContainer>
-							<Lottie
-								options={SparkleAnimationOptions}
-								height={100}
-								width={100}
-								isPaused={showClaimModal}
-								speed={0.8}
-							/>
-						</SparkleContainer>
-						<SunImage>
-							<Image
-								src='/images/union.svg'
-								height='115'
-								width='194'
-								alt='union'
-							/>
-						</SunImage>
-						<ClaimedContainer>
-							<ClaimedTitle>Congratulations!</ClaimedTitle>
-							<ClaimedSubtitleContainer>
-								<ClaimedSubtitleA>
-									You have successfully claimed{' '}
-									{formatWeiHelper(totalAmount.div(10))} GIV.{' '}
-									<AddGivButton
-										onClick={() =>
-											addGIVToken(
-												config.XDAI_NETWORK_NUMBER,
-											)
-										}
-									>
-										<Image
-											src='/images/icons/metamask.svg'
-											height='24'
-											width='24'
-											alt='Metamask logo.'
-										/>
-									</AddGivButton>
-								</ClaimedSubtitleA>
-								<ClaimedSubtitleB>
-									Plus you&apos;re getting an additional{' '}
-									<span style={{ color: '#FED670' }}>
-										{streamValue} GIV
-									</span>{' '}
-									per week.
-								</ClaimedSubtitleB>
-								<a
-									href='https://twitter.com/intent/tweet?text=The%20%23GIVeconomy%20is%20here!%20Excited%20to%20be%20part%20of%20the%20Future%20of%20Giving%20with%20$GIV%20%26%20%40givethio%20%23blockchain4good%20%23defi4good%20%23givethlove%20%23givdrop'
-									target='_blank'
-									rel='noreferrer'
-								>
-									<SocialButton>
-										share on twitter
-										<Image
-											src='/images/icons/twitter.svg'
-											height='15'
-											width='15'
-											alt='Twitter logo.'
-										/>
-									</SocialButton>
-								</a>
-								<a
-									href='https://swag.giveth.io/'
-									target='_blank'
-									rel='noreferrer'
-								>
-									<SocialButton>
-										claim your free swag
-										<Image
-											src='/images/icons/tshirt.svg'
-											height='15'
-											width='15'
-											alt='T shirt.'
-										/>
-									</SocialButton>
-								</a>
-								<a
-									href='https://discord.giveth.io/'
-									target='_blank'
-									rel='noreferrer'
-								>
-									<SocialButton>
-										join our discord
-										<Image
-											src='/images/icons/discord.svg'
-											height='15'
-											width='15'
-											alt='discord logo.'
-										/>
-									</SocialButton>
-								</a>
-								<Link href='/' passHref>
-									<a target='_blank' rel='noreferrer'>
-										<ExploreButton>
-											explore the giveconomy
-										</ExploreButton>
-									</a>
-								</Link>
-								<ClaimFromAnother
-									onClick={() => {
-										goFirstStep();
-										resetWallet();
-										setIsClaimed(false);
-									}}
-								>
-									Claim from another address!
-								</ClaimFromAnother>
-							</ClaimedSubtitleContainer>
-						</ClaimedContainer>
-					</>
-				) : (
-					<>
-						<ClaimHeader>
-							<Title as='h1' weight={700}>
-								Claim your GIV now!
-							</Title>
-							<Desc size='small' color={'#CABAFF'}>
-								Join the giving economy.
-							</Desc>
-						</ClaimHeader>
-						<Row alignItems={'center'} justifyContent={'center'}>
-							{/* <ClaimButton secondary onClick={onClaim}> */}
-							<ClaimButton
-								secondary
-								onClick={() => {
-									openHarvestModal();
-								}}
-							>
-								CLAIM {formatWeiHelper(totalAmount.div(10))} GIV
-							</ClaimButton>
-						</Row>
-						<Row alignItems={'center'} justifyContent={'center'}>
-							<MetamaskButton
-								onClick={() =>
-									addGIVToken(config.XDAI_NETWORK_NUMBER)
-								}
-							>
-								<Image
-									src='/images/metamask.png'
-									height='32'
-									width='215'
-									alt='Metamask button'
-								/>
-							</MetamaskButton>
-						</Row>
-					</>
-				)}
+				<ClaimHeader>
+					<Title as='h1' weight={700}>
+						Claim your GIV now!
+					</Title>
+					<Desc size='small' color={'#CABAFF'}>
+						Join the giving economy.
+					</Desc>
+				</ClaimHeader>
+				<Row alignItems={'center'} justifyContent={'center'}>
+					{/* <ClaimButton secondary onClick={onClaim}> */}
+					<ClaimButton
+						secondary
+						onClick={() => {
+							openHarvestModal();
+						}}
+					>
+						CLAIM {formatWeiHelper(totalAmount.div(10))} GIV
+					</ClaimButton>
+				</Row>
+				<Row alignItems={'center'} justifyContent={'center'}>
+					<MetamaskButton
+						onClick={() => addGIVToken(config.XDAI_NETWORK_NUMBER)}
+					>
+						<Image
+							src='/images/metamask.png'
+							height='32'
+							width='215'
+							alt='Metamask button'
+						/>
+					</MetamaskButton>
+				</Row>
 				{activeIndex === index && (
 					<>
 						<PreviousArrowButton onClick={goPreviousStep} />
