+++ src/components/cards/Congratulations.tsx
@@ -1,9 +1,83 @@
 import styled from 'styled-components';
 import { Button } from '../styled-components/Button';
-import { useContext } from 'react';
-import { UserContext } from '../../context/user.context';
+import { useContext, useEffect, useState } from 'react';
+import useUser, { UserContext } from '../../context/user.context';
 import { utils } from 'ethers';
-import { H2, Lead, P } from '@giveth/ui-design-system';
+import { Container, H2, Lead, P } from '@giveth/ui-design-system';
+import Lottie from 'react-lottie';
+import { formatWeiHelper } from '@/helpers/number';
+import { addGIVToken } from '@/lib/metamask';
+import Image from 'next/image';
+import SparkleBurstAnimation from '../../animations/sparkle-burst.json';
+import SparkleAnimation from '../../animations/sparkle.json';
+import config from '@/configuration';
+import Link from 'next/link';
+import { useTokenDistro } from '@/context/tokenDistro.context';
+import { ClaimViewContext } from '../views/claim/Claim.view';
+import { Row } from '../styled-components/Grid';
+
+const ClaimedContainer = styled.div`
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	position: relative;
+`;
+
+const SmileImage = styled.div`
+	position: absolute;
+	height: 0px;
+	left: -264px;
+	top: -92px;
+	@media only screen and (max-width: 990px) {
+		display: none;
+	}
+`;
+
+const ClaimedSubtitleA = styled(Row)`
+	gap: 12px;
+	margin-top: 24px;
+	position: relative;
+`;
+
+const AddGivButton = styled.div`
+	cursor: pointer;
+`;
+
+const ClaimedSubtitleB = styled(Lead)``;
+
+const SocialButtonsContainer = styled(Row)`
+	flex-direction: column;
+	gap: 12px;
+	margin: 32px 0;
+`;
+
+const SocialButton = styled(Button)`
+	font-family: 'Red Hat Text';
+	font-size: 14px;
+	font-weight: bold;
+	text-transform: uppercase;
+	background-color: transparent;
+	border: 2px solid white;
+	height: 50px;
+	width: 265px;
+	margin-top: 12px;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	gap: 4px;
+`;
+
+const ExploreButton = styled(SocialButton)`
+	background-color: #e1458d;
+	border: none;
+	width: 285px;
+`;
+
+const ClaimFromAnother = styled.div`
+	cursor: pointer;
+	color: #fed670;
+	margin-top: 16px;
+`;
 
 const CongratulationsView = styled.div`
 	min-height: 100vh;
@@ -12,73 +86,197 @@ const CongratulationsView = styled.div`
 	background-size: cover;
 `;
 
-const CongratulationsContainer = styled.div`
-	width: 700px;
-	margin: auto;
-	text-align: center;
+const CongratulationsContainer = styled(Container)`
+	min-height: 100vh;
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
 `;
 
-const CongHeader = styled.div`
+const CongHeader = styled(H2)`
 	text-align: center;
-	padding-top: 164px;
 	position: relative;
-	margin-bottom: 80px;
-	::before {
-		content: '';
-		background-image: url('/images/congratulation.png');
-		position: absolute;
-		width: 252px;
-		height: 226px;
-		bottom: 0;
-		left: -110px;
+`;
+
+const CongContent = styled(Row)`
+	flex-direction: column;
+	padding-left: 200px;
+	@media only screen and (max-width: 990px) {
+		padding-left: 0;
+		align-items: center;
 	}
 `;
 
-const Description = styled.div`
-	padding: 16px 10px 90px 164px;
-	text-align: left;
+const SparkleContainer = styled.div`
+	position: absolute;
+	right: -75px;
+	top: -48px;
 `;
 
-const Rate = styled(P)`
-	color: #fed670;
-	display: inline-block;
-	padding: 10px;
+const SparkleBurstContainer = styled.div`
+	position: absolute;
+	left: -44px;
+	top: -37px;
+	@media only screen and (max-width: 1360px) {
+	}
+	@media only screen and (max-width: 990px) {
+		display: none;
+	}
 `;
 
-const CongratsButton = styled(Button)`
-	width: 356px;
-	text-transform: uppercase;
-	font-size: 16px;
-	font-weight: 700;
-	line-height: 18px;
-	letter-spacing: 0.04em;
-	margin-top: 36px;
-`;
+const SparkleAnimationOptions = {
+	loop: false,
+	animationData: SparkleAnimation,
+	rendererSettings: {
+		preserveAspectRatio: 'xMidYMid slice',
+	},
+};
+
+const SparkleBurstAnimationOptions = {
+	loop: false,
+	animationData: SparkleBurstAnimation,
+	rendererSettings: {
+		preserveAspectRatio: 'xMidYMid slice',
+	},
+};
 
 export const CongratulationsCard = () => {
-	const { totalAmount } = useContext(UserContext);
+	const [streamValue, setStreamValue] = useState<string>('0');
+
+	const { tokenDistroHelper } = useTokenDistro();
+
+	const { goFirstStep } = useContext(ClaimViewContext);
+
+	const { userAddress, totalAmount, resetWallet } = useUser();
+
+	useEffect(() => {
+		setStreamValue(
+			formatWeiHelper(
+				tokenDistroHelper.getStreamPartTokenPerWeek(totalAmount),
+			),
+		);
+	}, [totalAmount, tokenDistroHelper]);
 	return (
 		<CongratulationsView>
 			<CongratulationsContainer>
-				<CongHeader>
-					<H2 as='h1' weight={700}>
-						Congratulations!
-					</H2>
-					<Description>
-						<Lead>
-							You have successfully claimed{' '}
-							{utils.formatEther(totalAmount.div(10))} GIV tokens.
-						</Lead>
-						<Lead>Add GIV to Metamask</Lead>
-					</Description>
+				<CongHeader weight={700}>
+					Congratulations!
+					<SparkleContainer>
+						<Lottie
+							options={SparkleAnimationOptions}
+							height={100}
+							width={100}
+							// isPaused={showClaimModal}
+							speed={0.8}
+						/>
+					</SparkleContainer>
 				</CongHeader>
-				<Lead>
-					And you&apos;re getting an additional
-					<Rate size='xlarge'>{10} GIV</Rate>
-					per second.
-				</Lead>
-				<Lead>Watch your GIVstream flow!</Lead>
-				<CongratsButton secondary>GIV me more</CongratsButton>
+				<CongContent>
+					<ClaimedSubtitleA>
+						<Lead>You have successfully claimed </Lead>
+						<Lead>
+							{formatWeiHelper(totalAmount.div(10))} GIV.{' '}
+						</Lead>
+						<AddGivButton
+							onClick={() =>
+								addGIVToken(config.XDAI_NETWORK_NUMBER)
+							}
+						>
+							<Image
+								src='/images/icons/metamask.svg'
+								height='24'
+								width='24'
+								alt='Metamask logo.'
+							/>
+						</AddGivButton>
+						<SmileImage>
+							<SparkleBurstContainer>
+								<Lottie
+									options={SparkleBurstAnimationOptions}
+									height={200}
+									width={200}
+									// isPaused={showClaimModal}
+									speed={0.8}
+								/>
+							</SparkleBurstContainer>
+							<Image
+								src='/images/congratulation.png'
+								height='226'
+								width='252'
+								alt='union'
+							/>
+						</SmileImage>
+					</ClaimedSubtitleA>
+					<SocialButtonsContainer>
+						<a
+							href='https:twitter.com/intent/tweet?text=The%20%23GIVeconomy%20is%20here!%20Excited%20to%20be%20part%20of%20the%20Future%20of%20Giving%20with%20$GIV%20%26%20%40givethio%20%23blockchain4good%20%23defi4good%20%23givethlove%20%23givdrop'
+							target='_blank'
+							rel='noreferrer'
+						>
+							<SocialButton>
+								share on twitter
+								<Image
+									src='/images/icons/twitter.svg'
+									height='15'
+									width='15'
+									alt='Twitter logo.'
+								/>
+							</SocialButton>
+						</a>
+						<a
+							href='https:swag.giveth.io/'
+							target='_blank'
+							rel='noreferrer'
+						>
+							<SocialButton>
+								claim your free swag
+								<Image
+									src='/images/icons/tshirt.svg'
+									height='15'
+									width='15'
+									alt='T shirt.'
+								/>
+							</SocialButton>
+						</a>
+						<a
+							href='https:discord.giveth.io/'
+							target='_blank'
+							rel='noreferrer'
+						>
+							<SocialButton>
+								join our discord
+								<Image
+									src='/images/icons/discord.svg'
+									height='15'
+									width='15'
+									alt='discord logo.'
+								/>
+							</SocialButton>
+						</a>
+					</SocialButtonsContainer>
+					<ClaimedSubtitleB>
+						Plus you&apos;re getting an additional{' '}
+						<span style={{ color: '#FED670' }}>
+							{streamValue} GIV
+						</span>{' '}
+						per week.
+					</ClaimedSubtitleB>
+					<ClaimedSubtitleB>
+						Watch your GIVstream flow!
+					</ClaimedSubtitleB>
+					<Link href='/givstream' passHref>
+						<ExploreButton>explore the giveconomy</ExploreButton>
+					</Link>
+					<ClaimFromAnother
+						onClick={() => {
+							goFirstStep();
+							resetWallet();
+						}}
+					>
+						Claim from another address!
+					</ClaimFromAnother>
+				</CongContent>
 			</CongratulationsContainer>
 		</CongratulationsView>
 	);
